Use BIGINT columns

+++ db.py
@@ -16,11 +16,11 @@ def init_db():
     with get_connection() as conn, conn.cursor() as cur:
         cur.execute(
             """CREATE TABLE IF NOT EXISTS rtp_history (
-            game_id INTEGER,
+            game_id BIGINT,
             name TEXT,
             provider TEXT,
             rtp REAL,
-            extra INTEGER,
+            extra BIGINT,
             timestamp TIMESTAMPTZ DEFAULT CURRENT_TIMESTAMP
         )"""
         )
